+++ app/src/main/java/com/richard/earthquake/app/presantation/controller/v1/authorized/EarthquakeController.java
@@ -36,7 +36,7 @@ public class EarthquakeController {
     }
 
     @GetMapping("/all")
-    public Mono<ResponseEntity<List<Earthquake>>> findAllEarthquakesForUser(@RequestParam Integer pageNumber, @RequestParam Integer elementPerPage, @RequestParam(required = false) Short countryKey, @RequestParam(required = false) Integer mag) {
+    public Mono<ResponseEntity<List<Earthquake>>> findAllEarthquakesForUser(@RequestParam Integer pageNumber, @RequestParam Integer elementPerPage, @RequestParam(required = false) Short countryKey, @RequestParam(required = false) Double mag) {
         return earthquakeService.getAllFilteredEarthquake(PageRequest.of(pageNumber, elementPerPage), countryKey, mag,false)
                 .collectList()
                 .map(ResponseEntity::ok)
